hoist the cache to the module level
During debugging, I noticed that every time the PostCSS code runs, the
full plugin re-runs. This means that caches are thrown away. I'm
currently not sure if this is a test specific issue or an actual issue.

That said, the integration tests do run postcss-cli for example directly
so I think this is a real issue.

This change should be safe since the cache is unique per input file
(which it was before as well). Worst case is that we use a bit more
memory.