Collect all changes and apply them in one go
This makes sure that we’re always slicing the original string and not a changing string. This means that we can traverse the candidate replacements forward through the string rather than slicing and applying them in reverse order. A little extra bookkeeping is required but produces a decent speedup in v8.